requirements.txt
streamlit>=1.35.0
cryptography>=42.0.0
pycryptodome>=3.20.0
plotly>=5.18.0
pycipher>=0.9.0
secretpy>=0.11.0
